updated file structure with suggested routes and naming conventions
